tests: Add test for a max free entries limit
